Move ZendeskManager.kt identity configuration to ZendeskSettings.kt
